database tx traits
