+++ crates/db/Cargo.toml
@@ -10,6 +10,7 @@ description = "Staged syncing primitives used in reth."
 [dependencies]
 # reth
 reth-primitives = { path = "../primitives" }
+reth-interfaces = { path = "../interfaces" }
 
 # misc
 bytes = "1.2.1"
