+++ crates/interfaces/src/db/error.rs
@@ -1,0 +1,4 @@
+
+/// Database Error
+#[derive(Debug)]
+pub enum KVError {}
